requirements.examples.txt
torchvision>=0.3
visdom>=0.1.8.8
paramiko
boto3
tqdm
